Simplify compass process
