+++ src/main/java/de/tum/in/www1/artemis/service/compass/CompassCalculationEngine.java
@@ -6,15 +6,14 @@ import static de.tum.in.www1.artemis.service.compass.utils.CompassConfiguration.
 import java.io.IOException;
 import java.time.LocalDateTime;
 import java.util.*;
-import java.util.stream.Collectors;
 
 import org.hibernate.Hibernate;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.hazelcast.core.HazelcastInstance;
+import com.hazelcast.cp.IAtomicReference;
 
 import de.tum.in.www1.artemis.domain.Feedback;
 import de.tum.in.www1.artemis.domain.Result;
@@ -23,10 +22,10 @@ import de.tum.in.www1.artemis.domain.enumeration.AssessmentType;
 import de.tum.in.www1.artemis.domain.enumeration.FeedbackType;
 import de.tum.in.www1.artemis.domain.modeling.ModelingSubmission;
 import de.tum.in.www1.artemis.service.compass.assessment.CompassResult;
-import de.tum.in.www1.artemis.service.compass.assessment.Score;
 import de.tum.in.www1.artemis.service.compass.assessment.SimilaritySetAssessment;
 import de.tum.in.www1.artemis.service.compass.controller.*;
 import de.tum.in.www1.artemis.service.compass.grade.Grade;
+import de.tum.in.www1.artemis.service.compass.statistics.CompassStatistics;
 import de.tum.in.www1.artemis.service.compass.umlmodel.UMLDiagram;
 import de.tum.in.www1.artemis.service.compass.umlmodel.UMLElement;
 import de.tum.in.www1.artemis.service.compass.umlmodel.classdiagram.*;
@@ -41,14 +40,17 @@ public class CompassCalculationEngine {
 
     private ModelSelector modelSelector;
 
-    private LocalDateTime lastUsed;
+    private IAtomicReference<LocalDateTime> lastUsed;
+
+    private CompassStatistics compassStatistics;
 
     CompassCalculationEngine(Long exerciseId, Set<ModelingSubmission> modelingSubmissions, HazelcastInstance hazelcastInstance) {
-        lastUsed = LocalDateTime.now();
+        lastUsed = hazelcastInstance.getCPSubsystem().getAtomicReference("lastUsed - " + exerciseId);
+        lastUsed.set(LocalDateTime.now());
         modelIndex = new ModelIndex(exerciseId, hazelcastInstance);
         automaticAssessmentController = new AutomaticAssessmentController(exerciseId, hazelcastInstance);
-        modelSelector = new ModelSelector(automaticAssessmentController);
-
+        modelSelector = new ModelSelector(automaticAssessmentController, hazelcastInstance, exerciseId);
+        compassStatistics = new CompassStatistics(modelIndex, automaticAssessmentController);
         for (Submission submission : modelingSubmissions) {
             // We have to unproxy here as sometimes the Submission is a Hibernate proxy resulting in a cast exception
             // when iterating over the ModelingSubmissions directly (i.e. for (ModelingSubmission submission : submissions)).
@@ -57,24 +59,10 @@ public class CompassCalculationEngine {
             String model = modelingSubmission.getModel();
             if (model != null) {
                 buildModel(modelingSubmission);
-
-                if (hasCompletedManualAssessment(modelingSubmission)) {
-                    addManualAssessmentForSubmission(modelingSubmission);
-                }
+                addManualAssessmentForSubmissionIfExists(modelingSubmission);
             }
         }
         assessModelsAutomatically();
-    }
-
-    /**
-     * Checks if the given modeling submission already has a completed manual assessment. The assessment is completed if the submission has a result with a completion date.
-     *
-     * @param modelingSubmission the modeling submission to check
-     * @return true if the submission already has a completed manual assessment, false otherwise
-     */
-    private boolean hasCompletedManualAssessment(ModelingSubmission modelingSubmission) {
-        return modelingSubmission.getLatestResult() != null && modelingSubmission.getLatestResult().getCompletionDate() != null
-                && modelingSubmission.getLatestResult().getAssessmentType().equals(AssessmentType.MANUAL);
     }
 
     /**
@@ -86,8 +74,12 @@ public class CompassCalculationEngine {
      */
     private void buildModel(ModelingSubmission modelingSubmission) {
         if (modelingSubmission.getModel() != null) {
-            buildModel(modelingSubmission.getId(), parseString(modelingSubmission.getModel()).getAsJsonObject());
+            buildModel(modelingSubmission.getId(), modelingSubmission.getModel());
         }
+    }
+
+    private void buildModel(long modelSubmissionId, String model) {
+        buildModel(modelSubmissionId, parseString(model).getAsJsonObject());
     }
 
     /**
@@ -108,16 +100,30 @@ public class CompassCalculationEngine {
         }
     }
 
+    public void buildPendingModels() {
+        for (Map.Entry<Long, String> modelEntry : modelIndex.getPendingEntries()) {
+            buildModel(modelEntry.getKey(), modelEntry.getValue());
+        }
+    }
+
     /**
      * Adds the manual assessment of the given submission to Compass so that it can be used for automatic assessments. Additionally, it marks the submission as assessed, i.e. the
      * submission is not considered when providing a submission for manual assessment to the client.
      *
      * @param submission the submission for which the manual assessment is added
      */
-    private void addManualAssessmentForSubmission(ModelingSubmission submission) {
-        UMLDiagram model = modelIndex.getModelMap().get(submission.getId());
+    private void addManualAssessmentForSubmissionIfExists(ModelingSubmission submission) {
 
-        if (model == null || submission.getLatestResult() == null || submission.getLatestResult().getCompletionDate() == null) {
+        Boolean hasManualAssessment = submission.getLatestResult() != null && submission.getLatestResult().getCompletionDate() != null
+                && submission.getLatestResult().getAssessmentType().equals(AssessmentType.MANUAL);
+
+        if (!hasManualAssessment) {
+            return;
+        }
+
+        UMLDiagram model = modelIndex.getModel(submission.getId());
+
+        if (model == null) {
             log.error("Could not build assessment for submission {}", submission.getId());
             return;
         }
@@ -126,24 +132,6 @@ public class CompassCalculationEngine {
 
         modelSelector.removeModelWaitingForAssessment(submission.getId());
         modelSelector.addAlreadyHandledModel(submission.getId());
-    }
-
-    protected Collection<UMLDiagram> getUmlModelCollection() {
-        return modelIndex.getModelCollection();
-    }
-
-    protected Map<Long, UMLDiagram> getModelMap() {
-        return modelIndex.getModelMap();
-    }
-
-    @SuppressWarnings("unused")
-    private double getTotalCoverage() {
-        return automaticAssessmentController.getTotalCoverage();
-    }
-
-    @SuppressWarnings("unused")
-    private double getTotalConfidence() {
-        return automaticAssessmentController.getTotalConfidence();
     }
 
     private void assessModelsAutomatically() {
@@ -158,7 +146,7 @@ public class CompassCalculationEngine {
      * @return the ids of the next optimal modeling submissions, or an empty list if there are no unhandled submissions
      */
     public List<Long> getNextOptimalModels(int numberOfModels) {
-        lastUsed = LocalDateTime.now();
+        lastUsed.set(LocalDateTime.now());
         return modelSelector.selectNextModels(modelIndex, numberOfModels);
     }
 
@@ -169,12 +157,13 @@ public class CompassCalculationEngine {
      * @return the assessment result for the model
      */
     public Grade getGradeForModel(long modelSubmissionId) {
-        lastUsed = LocalDateTime.now();
-        if (!modelIndex.getModelMap().containsKey(modelSubmissionId)) {
+        lastUsed.set(LocalDateTime.now());
+
+        UMLDiagram model = modelIndex.getModel(modelSubmissionId);
+        if (model == null) {
             return null;
         }
 
-        UMLDiagram model = modelIndex.getModelMap().get(modelSubmissionId);
         CompassResult compassResult = automaticAssessmentController.getLastAssessmentCompassResult(modelSubmissionId);
 
         if (compassResult == null) {
@@ -184,7 +173,7 @@ public class CompassCalculationEngine {
     }
 
     public Collection<Long> getModelIds() {
-        return modelIndex.getModelMap().keySet();
+        return modelIndex.getModelIds();
     }
 
     /**
@@ -194,7 +183,7 @@ public class CompassCalculationEngine {
      * @param assessedModelSubmissionId  the id of the corresponding model
      */
     public void notifyNewAssessment(List<Feedback> modelingAssessment, long assessedModelSubmissionId) {
-        lastUsed = LocalDateTime.now();
+        lastUsed.set(LocalDateTime.now());
         modelSelector.addAlreadyHandledModel(assessedModelSubmissionId);
         UMLDiagram model = modelIndex.getModel(assessedModelSubmissionId);
         if (model == null) {
@@ -213,28 +202,13 @@ public class CompassCalculationEngine {
      * @param modelId the id of the new model
      */
     public void notifyNewModel(String model, long modelId) {
-        lastUsed = LocalDateTime.now();
+        lastUsed.set(LocalDateTime.now());
         // Do not add models that might already exist
-        if (modelIndex.getModelMap().containsKey(modelId)) {
+        if (modelIndex.modelExists(modelId)) {
             return;
         }
         if (model != null) {
-            JsonElement jsonElement = parseString(model);
-            if (jsonElement != null) {
-                buildModel(modelId, jsonElement.getAsJsonObject());
-            }
-        }
-    }
-
-    /**
-     * Add a new model
-     *
-     * @param submissions the list of submissions to get models and ids from
-     */
-    public void notifyNewModels(List<ModelingSubmission> submissions) {
-        lastUsed = LocalDateTime.now();
-        for (ModelingSubmission submission : submissions) {
-            notifyNewModel(submission.getModel(), submission.getId());
+            modelIndex.addPendingModel(modelId, model);
         }
     }
 
@@ -242,7 +216,7 @@ public class CompassCalculationEngine {
      * @return the time when the engine has been used last
      */
     public LocalDateTime getLastUsedAt() {
-        return lastUsed;
+        return lastUsed.get();
     }
 
     /**
@@ -291,7 +265,7 @@ public class CompassCalculationEngine {
      * @return the list of Feedback items generated from the Grade
      */
     public List<Feedback> convertToFeedback(Grade grade, long modelId, Result result) {
-        UMLDiagram model = this.modelIndex.getModelMap().get(modelId);
+        UMLDiagram model = this.modelIndex.getModel(modelId);
 
         if (model == null || grade == null || grade.getJsonIdPointsMapping() == null) {
             return new ArrayList<>();
@@ -312,7 +286,7 @@ public class CompassCalculationEngine {
 
             // Get the confidence for this element of the model. If the confidence is less than the configured threshold, no automatic feedback will be created for this element
             // and the loop will continue with the next model element.
-            double elementConfidence = getConfidenceForElement(jsonElementID, modelId);
+            double elementConfidence = getConfidenceForElement(jsonElementID);
             if (elementConfidence < ELEMENT_CONFIDENCE_THRESHOLD) {
                 log.debug("Confidence {} of element {} is smaller than configured confidence threshold {}", elementConfidence, jsonElementID, ELEMENT_CONFIDENCE_THRESHOLD);
                 continue;
@@ -344,76 +318,19 @@ public class CompassCalculationEngine {
     }
 
     /**
-     * format: uniqueElements [{id} name apollonId conflicts] numberModels numberConflicts totalConfidence totalCoverage models [{id} confidence coverage conflicts]
-     *
-     * @return statistics about the UML model
-     */
-    // TODO: I don't think we should expose JSONObject to this internal server class. It would be better to return Java objects here
-    public JsonObject getStatistics() {
-        JsonObject jsonObject = new JsonObject();
-
-        JsonObject uniqueElements = new JsonObject();
-        int numberOfConflicts = 0;
-        for (UMLElement umlElement : modelIndex.getUniqueElements()) {
-            JsonObject uniqueElement = new JsonObject();
-            uniqueElement.addProperty("name", umlElement.toString());
-            uniqueElement.addProperty("apollonId", umlElement.getJSONElementID());
-            boolean hasConflict = hasConflict(umlElement.getSimilarityID());
-            if (hasConflict) {
-                numberOfConflicts++;
-            }
-            uniqueElement.addProperty("conflicts", hasConflict);
-            uniqueElements.add(String.valueOf(umlElement.getSimilarityID()), uniqueElement);
-        }
-        jsonObject.add("uniqueElements", uniqueElements);
-
-        jsonObject.addProperty("numberModels", modelIndex.getModelCollection().size());
-        jsonObject.addProperty("numberConflicts", numberOfConflicts);
-        jsonObject.addProperty("totalConfidence", getTotalConfidence());
-        jsonObject.addProperty("totalCoverage", getTotalCoverage());
-
-        JsonObject models = new JsonObject();
-        for (Map.Entry<Long, UMLDiagram> modelEntry : getModelMap().entrySet()) {
-            JsonObject model = new JsonObject();
-            model.addProperty("coverage", automaticAssessmentController.getLastAssessmentCoverage(modelEntry.getValue().getModelSubmissionId()));
-            model.addProperty("confidence", automaticAssessmentController.getLastAssessmentConfidence(modelEntry.getValue().getModelSubmissionId()));
-            int numberOfModelConflicts = 0;
-            List<UMLElement> elements = new ArrayList<>(modelEntry.getValue().getAllModelElements());
-            for (UMLElement element : elements) {
-                boolean modelHasConflict = hasConflict(element.getSimilarityID());
-                if (modelHasConflict) {
-                    numberOfModelConflicts++;
-                }
-            }
-            model.addProperty("conflicts", numberOfModelConflicts);
-            model.addProperty("elements", elements.size());
-            model.addProperty("classes", elements.stream().filter(umlElement -> umlElement instanceof UMLClass).count());
-            model.addProperty("attributes", elements.stream().filter(umlElement -> umlElement instanceof UMLAttribute).count());
-            model.addProperty("methods", elements.stream().filter(umlElement -> umlElement instanceof UMLMethod).count());
-            model.addProperty("associations", elements.stream().filter(umlElement -> umlElement instanceof UMLRelationship).count());
-            models.add(modelEntry.getKey().toString(), model);
-        }
-        jsonObject.add("models", models);
-
-        return jsonObject;
-    }
-
-    /**
      * Get the confidence for a specific model element with the given id. The confidence is the percentage indicating how certain Compass is about the assessment of the given model
      * element. If it is smaller than a configured threshold, the element with the given id will not be assessed automatically.
      *
      * @param elementId    the id of the model element
-     * @param submissionId the id of the submission that contains the corresponding model
      * @return the confidence for the model element with the given id
      */
-    public double getConfidenceForElement(String elementId, long submissionId) {
-        UMLDiagram model = modelIndex.getModel(submissionId);
-        UMLElement element = model.getElementByJSONID(elementId);
-        if (element == null) {
+    public double getConfidenceForElement(String elementId) {
+        Integer similarityId = modelIndex.getSimilarityId(elementId);
+        if (similarityId == null) {
             return 0.0;
         }
 
-        Optional<SimilaritySetAssessment> optionalAssessment = automaticAssessmentController.getAssessmentForSimilaritySet(element.getSimilarityID());
+        Optional<SimilaritySetAssessment> optionalAssessment = automaticAssessmentController.getAssessmentForSimilaritySet(similarityId);
         return optionalAssessment.map(assessment -> assessment.getScore().getConfidence()).orElse(0.0);
     }
 
@@ -428,18 +345,6 @@ public class CompassCalculationEngine {
         automaticAssessmentController.addFeedbacksToSimilaritySet(modelingAssessment, model);
     }
 
-    private boolean hasConflict(int elementId) {
-        Optional<SimilaritySetAssessment> optionalAssessment = automaticAssessmentController.getAssessmentForSimilaritySet(elementId);
-
-        if (optionalAssessment.isEmpty() || optionalAssessment.get().getFeedbackList() == null || optionalAssessment.get().getFeedbackList().isEmpty()) {
-            return false;
-        }
-
-        List<Feedback> feedbackList = optionalAssessment.get().getFeedbackList();
-        // if not all feedback entries have the same score as the first feedback, i.e. not all scores are the same, there is a conflict in the assessment
-        return !feedbackList.stream().allMatch(feedback -> feedback.getCredits().equals(feedbackList.get(0).getCredits()));
-    }
-
     /**
      * Used for internal analysis of modeling data. Do not remove, usage is commented out due to performance reasons.
      *
@@ -447,200 +352,23 @@ public class CompassCalculationEngine {
      * @param finishedResults the list of finished results, i.e. results for which assessor and completion date is not null
      */
     public void printStatistic(long exerciseId, List<Result> finishedResults) {
-        log.info("Statistics for exercise {}\n\n\n", exerciseId);
-
-        long totalNumberOfFeedback = 0;
-        long totalNumberOfAutomaticFeedback = 0;
-        long totalNumberOfAdaptedFeedback = 0;
-        long totalNumberOfManualFeedback = 0;
-
-        long numberOfAssessedClasses = 0;
-        long numberOfAssessedAttrbutes = 0;
-        long numberOfAssessedMethods = 0;
-        long numberOfAssessedRelationships = 0;
-        long numberOfAssessedPackages = 0;
-
-        long totalLengthOfFeedback = 0;
-        long totalLengthOfPositiveFeedback = 0;
-        long totalNumberOfPositiveFeedbackItems = 0;
-        long totalLengthOfNegativeFeedback = 0;
-        long totalNumberOfNegativeFeedbackItems = 0;
-        long totalLengthOfNeutralFeedback = 0;
-        long totalNumberOfNeutralFeedbackItems = 0;
-
-        for (Result result : finishedResults) {
-            List<Feedback> referenceFeedback = result.getFeedbacks().stream().filter(feedback -> feedback.getReference() != null).collect(Collectors.toList());
-
-            totalNumberOfFeedback += referenceFeedback.size();
-            totalNumberOfAutomaticFeedback += referenceFeedback.stream().filter(feedback -> feedback.getType() == FeedbackType.AUTOMATIC).count();
-            totalNumberOfAdaptedFeedback += referenceFeedback.stream().filter(feedback -> feedback.getType() == FeedbackType.AUTOMATIC_ADAPTED).count();
-            totalNumberOfManualFeedback += referenceFeedback.stream().filter(feedback -> feedback.getType() == FeedbackType.MANUAL).count();
-
-            numberOfAssessedClasses += referenceFeedback.stream()
-                    .filter(feedback -> feedback.getReferenceElementType().equals("Class") || feedback.getReferenceElementType().equals("AbstractClass")
-                            || feedback.getReferenceElementType().equals("Interface") || feedback.getReferenceElementType().equals("Enumeration"))
-                    .count();
-            numberOfAssessedAttrbutes += referenceFeedback.stream().filter(feedback -> feedback.getReferenceElementType().equals("ClassAttribute")).count();
-            numberOfAssessedMethods += referenceFeedback.stream().filter(feedback -> feedback.getReferenceElementType().equals("ClassMethod")).count();
-            numberOfAssessedRelationships += referenceFeedback.stream()
-                    .filter(feedback -> feedback.getReferenceElementType().equals("ClassBidirectional") || feedback.getReferenceElementType().equals("ClassUnidirectional")
-                            || feedback.getReferenceElementType().equals("ClassAggregation") || feedback.getReferenceElementType().equals("ClassInheritance")
-                            || feedback.getReferenceElementType().equals("ClassDependency") || feedback.getReferenceElementType().equals("ClassComposition")
-                            || feedback.getReferenceElementType().equals("ClassRealization"))
-                    .count();
-            numberOfAssessedPackages += referenceFeedback.stream().filter(feedback -> feedback.getReferenceElementType().equals("Package")).count();
-
-            for (Feedback feedback : referenceFeedback) {
-                int feedbackLength = 0;
-
-                if (feedback.getText() != null) {
-                    feedbackLength = feedback.getText().length();
-                }
-
-                totalLengthOfFeedback += feedbackLength;
-
-                if (feedback.getCredits() > 0) {
-                    totalLengthOfPositiveFeedback += feedbackLength;
-                    totalNumberOfPositiveFeedbackItems++;
-                }
-                else if (feedback.getCredits() == 0) {
-                    totalLengthOfNeutralFeedback += feedbackLength;
-                    totalNumberOfNeutralFeedbackItems++;
-                }
-                else if (feedback.getCredits() < 0) {
-                    totalLengthOfNegativeFeedback += feedbackLength;
-                    totalNumberOfNegativeFeedbackItems++;
-                }
-            }
-        }
-
-        long numberOfModels = modelIndex.getModelCollection().size();
-        Map<UMLElement, Integer> modelElementMapping = modelIndex.getElementSimilarityMap();
-        long numberOfModelElements = modelElementMapping.size();
-        long numberOfClasses = 0;
-        long numberOfAttrbutes = 0;
-        long numberOfMethods = 0;
-        long numberOfRelationships = 0;
-        long numberOfPackages = 0;
-
-        for (UMLElement element : modelElementMapping.keySet()) {
-            if (element instanceof UMLClass) {
-                numberOfClasses++;
-            }
-            else if (element instanceof UMLAttribute) {
-                numberOfAttrbutes++;
-            }
-            else if (element instanceof UMLMethod) {
-                numberOfMethods++;
-            }
-            else if (element instanceof UMLRelationship) {
-                numberOfRelationships++;
-            }
-            else if (element instanceof UMLPackage) {
-                numberOfPackages++;
-            }
-        }
-
-        // General information
-        log.info("################################################## General information ##################################################\n");
-
-        log.info("Number of models: {}\n", numberOfModels);
-        log.info("Number of model elements: {}\n", numberOfModelElements);
-        log.info("Number of classes: {}\n", numberOfClasses);
-        log.info("Number of attributes: {}\n", numberOfAttrbutes);
-        log.info("Number of methods: {}\n", numberOfMethods);
-        log.info("Number of relationships: {}\n", numberOfRelationships);
-        log.info("Number of packages: {}\n", numberOfPackages);
-        double elementsPerModel = numberOfModelElements * 1.0 / numberOfModels;
-        log.info("Average number of elements per model: {}\n", elementsPerModel);
-
-        log.info("Number of assessed models: {}\n", finishedResults.size());
-        log.info("Number of assessed model elements: {}\n", totalNumberOfFeedback);
-        log.info("Number of assessed classes: {} ({}%)\n", numberOfAssessedClasses, Math.round(numberOfAssessedClasses * 10000.0 / numberOfClasses) / 100.0);
-        log.info("Number of assessed attributes: {} ({}%)\n", numberOfAssessedAttrbutes, Math.round(numberOfAssessedAttrbutes * 10000.0 / numberOfAttrbutes) / 100.0);
-        log.info("Number of assessed methods: {} ({}%)\n", numberOfAssessedMethods, Math.round(numberOfAssessedMethods * 10000.0 / numberOfMethods) / 100.0);
-        log.info("Number of assessed relationships: {} ({}%)\n", numberOfAssessedRelationships,
-                Math.round(numberOfAssessedRelationships * 10000.0 / numberOfRelationships) / 100.0);
-        log.info("Number of assessed packages: {} ({}%)\n", numberOfAssessedPackages, Math.round(numberOfAssessedPackages * 10000.0 / numberOfPackages) / 100.0);
-        double feedbackPerAssessment = totalNumberOfFeedback * 1.0 / finishedResults.size();
-        log.info("Average number of feedback elements per assessment: {}\n\n\n", feedbackPerAssessment);
-
-        // Feedback type
-        log.info("################################################## Feedback type ##################################################\n");
-
-        log.info("Automatic feedback: {} ({}%)\n", totalNumberOfAutomaticFeedback, Math.round(totalNumberOfAutomaticFeedback * 10000.0 / totalNumberOfFeedback) / 100.0);
-        log.info("Adapted feedback: {} ({}%)\n", totalNumberOfAdaptedFeedback, Math.round(totalNumberOfAdaptedFeedback * 10000.0 / totalNumberOfFeedback) / 100.0);
-        log.info("Manual feedback: {} ({}%)\n", totalNumberOfManualFeedback, Math.round(totalNumberOfManualFeedback * 10000.0 / totalNumberOfFeedback) / 100.0);
-        log.info("Amount of automatic feedback that was adapted: {}%\n\n\n",
-                Math.round(totalNumberOfAdaptedFeedback * 10000.0 / (totalNumberOfAutomaticFeedback + totalNumberOfAdaptedFeedback)) / 100.0);
-
-        // Feedback length
-        log.info("################################################## Feedback length ##################################################\n");
-
-        log.info("Total amount of feedback: {}\n", totalNumberOfFeedback);
-        log.info("Average length of feedback: {}\n", totalLengthOfFeedback * 1.0 / totalNumberOfFeedback);
-        log.info("Total amount of positive feedback: {}\n", totalNumberOfPositiveFeedbackItems);
-        log.info("Average length of positive feedback: {}\n", totalLengthOfPositiveFeedback * 1.0 / totalNumberOfPositiveFeedbackItems);
-        log.info("Total amount of neutral feedback: {}\n", totalNumberOfNeutralFeedbackItems);
-        log.info("Average length of neutral feedback: {}\n", totalLengthOfNeutralFeedback * 1.0 / totalNumberOfNeutralFeedbackItems);
-        log.info("Total amount of negative feedback: {}\n", totalNumberOfNegativeFeedbackItems);
-        log.info("Average length of negative feedback: {}\n\n\n", totalLengthOfNegativeFeedback * 1.0 / totalNumberOfNegativeFeedbackItems);
-
-        // Similarity sets
-        log.info("################################################## Similarity sets ##################################################\n");
-
-        // Note, that these two value refer to all similarity sets that have an assessment, i.e. it is not the total number as it excludes the sets without assessments. This might
-        // distort the analysis values below.
-        long numberOfSimilaritySets = automaticAssessmentController.getAssessmentMap().size();
-        long numberOfElementsInSimilaritySets = 0;
-
-        long numberOfSimilaritySetsPositiveScore = 0;
-        long numberOfSimilaritySetsPositiveScoreRegardingConfidence = 0;
-
-        for (SimilaritySetAssessment similaritySetAssessment : automaticAssessmentController.getAssessmentMap().values()) {
-            numberOfElementsInSimilaritySets += similaritySetAssessment.getFeedbackList().size();
-
-            Score score = similaritySetAssessment.getScore();
-            if (score.getPoints() > 0) {
-                numberOfSimilaritySetsPositiveScore += 1;
-                if (score.getConfidence() >= 0.8) {
-                    numberOfSimilaritySetsPositiveScoreRegardingConfidence += 1;
-                }
-            }
-        }
-
-        log.info("Number of unique elements (without context) of submitted models: {}\n", modelIndex.getNumberOfUniqueElements());
-        log.info("Number of similarity sets (including context) of assessed models: {}\n", numberOfSimilaritySets);
-        log.info("Average number of elements per similarity set: {}\n", numberOfElementsInSimilaritySets * 1.0 / numberOfSimilaritySets);
-        // The optimal correction effort describes the maximum amount of model elements that tutors would have to assess in an optimal scenario
-        log.info("Optimal correction effort (# similarity sets / # model elements): {}\n", numberOfSimilaritySets * 1.0 / numberOfElementsInSimilaritySets);
-
-        log.info("Number of similarity sets with positive score: {}\n", numberOfSimilaritySetsPositiveScore);
-        log.info("Number of similarity sets with positive score and confidence at least 80%: {}\n\n\n", numberOfSimilaritySetsPositiveScoreRegardingConfidence);
-
-        // Variability index
-        log.info("################################################## Variability index ##################################################\n");
-
-        log.info("Variability index #1 (positive score): {}\n", numberOfSimilaritySetsPositiveScore / elementsPerModel);
-        log.info("Variability index #2 (positive score and confidence >= 80%): {}\n", numberOfSimilaritySetsPositiveScoreRegardingConfidence / elementsPerModel);
-        log.info("Variability index #3 (based on \"all\" similarity sets): {}\n", numberOfSimilaritySets / elementsPerModel);
-
-        log.info("Normalized variability index #1 (positive score): {}\n", (numberOfSimilaritySetsPositiveScore - elementsPerModel) / (numberOfModelElements - elementsPerModel));
-        log.info("Normalized variability index #2 (positive score and confidence >= 80%): {}\n",
-                (numberOfSimilaritySetsPositiveScoreRegardingConfidence - elementsPerModel) / (numberOfModelElements - elementsPerModel));
-        log.info("Normalized variability index #3 (based on \"all\" similarity sets): {}\n",
-                (numberOfSimilaritySets - elementsPerModel) / (numberOfModelElements - elementsPerModel));
-
-        // Alternative calculation of the variability index considering the average feedback items per assessment instead of the average elements per model
-        log.info("Alternative variability index #1 (positive score): {}\n", numberOfSimilaritySetsPositiveScore / feedbackPerAssessment);
-        log.info("Alternative variability index #2 (positive score and confidence >= 80%): {}\n", numberOfSimilaritySetsPositiveScoreRegardingConfidence / feedbackPerAssessment);
-        log.info("Alternative variability index #3 (based on \"all\" similarity sets): {}\n", numberOfSimilaritySets / feedbackPerAssessment);
-
-        log.info("Normalized alternative variability index #1 (positive score): {}\n",
-                (numberOfSimilaritySetsPositiveScore - feedbackPerAssessment) / (totalNumberOfFeedback - feedbackPerAssessment));
-        log.info("Normalized alternative variability index #2 (positive score and confidence >= 80%): {}\n",
-                (numberOfSimilaritySetsPositiveScoreRegardingConfidence - feedbackPerAssessment) / (totalNumberOfFeedback - feedbackPerAssessment));
-        log.info("Normalized alternative variability index #3 (based on \"all\" similarity sets): {}\n",
-                (numberOfSimilaritySets - feedbackPerAssessment) / (totalNumberOfFeedback - feedbackPerAssessment));
+        compassStatistics.printStatistic(exerciseId, finishedResults);
     }
+
+    /**
+     * format: uniqueElements [{id} name apollonId conflicts] numberModels numberConflicts totalConfidence totalCoverage models [{id} confidence coverage conflicts]
+     *
+     * @return statistics about the UML model
+     */
+    // TODO: I don't think we should expose JSONObject to this internal server class. It would be better to return Java objects here
+    public JsonObject getStatistics() {
+        return compassStatistics.getStatistics();
+    }
+
+    public void destroy() {
+        modelIndex.destroy();
+        automaticAssessmentController.destroy();
+        modelSelector.destroy();
+    }
+
 }
