+++ src/main/java/de/tum/in/www1/artemis/service/compass/controller/AutomaticAssessmentController.java
@@ -9,6 +9,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.hazelcast.core.HazelcastInstance;
+import com.hazelcast.map.IMap;
 
 import de.tum.in.www1.artemis.domain.Feedback;
 import de.tum.in.www1.artemis.service.compass.assessment.CompassResult;
@@ -21,9 +22,9 @@ public class AutomaticAssessmentController {
 
     private final Logger log = LoggerFactory.getLogger(AutomaticAssessmentController.class);
 
-    private Map<Integer, SimilaritySetAssessment> similarityIdAssessmentMapping;
+    private IMap<Integer, SimilaritySetAssessment> similarityIdAssessmentMapping;
 
-    private Map<Long, CompassResult> lastAssessmentResultMapping;
+    private IMap<Long, CompassResult> lastAssessmentResultMapping;
 
     HazelcastInstance hazelcastInstance;
 
@@ -260,5 +261,10 @@ public class AutomaticAssessmentController {
         }
 
         return getLastAssessmentCompassResult(submissionId).getCoverage();
+    }
+
+    public void destroy() {
+        similarityIdAssessmentMapping.destroy();
+        lastAssessmentResultMapping.destroy();
     }
 }
