+++ src/main/webapp/app/exercises/modeling/manage/modeling-exercise-update.component.ts
@@ -43,7 +43,7 @@ export class ModelingExerciseUpdateComponent implements OnInit {
     isImport: boolean;
     isExamMode: boolean;
     // TODO: Melih Oezbeyli(iozbeyli) Make it true after compass hazelcast problem is resolved
-    semiAutomaticAssessmentAvailable = false;
+    semiAutomaticAssessmentAvailable = true;
 
     constructor(
         private jhiAlertService: JhiAlertService,
